==> analysis/integration/mc_rna/broad_tenx/Summary/03.make_integration_summary.ipynb <==
import pandas as pd
import seaborn as sns
import matplotlib as mpl
import matplotlib.pyplot as plt
import numpy as np
import pathlib

import anndata

from ALLCools.mcds import MCDS
from ALLCools.plot import *
from wmb import *

adata = anndata.read_h5ad('../L1/Neuron/final_with_coords.h5ad')

tsne_df = pd.DataFrame(adata.obsm['X_tsne'], index=adata.obs_names, columns=['tsne_0', 'tsne_1'])
umap_df = pd.DataFrame(adata.obsm['X_umap'], index=adata.obs_names, columns=['umap_0', 'umap_1'])

ref_group = pd.read_csv('../L1/Neuron/mc_integration_group.csv.gz', index_col=0).squeeze()
query_group = pd.read_csv('../L1/Neuron/rna_integration_group.csv.gz', index_col=0).squeeze()
adata.obs['InteGroup'] = pd.concat([ref_group, query_group])

cluster_df = adata.obs[['leiden', 'InteGroup', 'Modality']]
cluster_df = cluster_df.rename(columns={'leiden': 'CoCluster'})

total_result = pd.concat([cluster_df, tsne_df, umap_df], axis=1)
total_result.columns = 'L1_' + total_result.columns

total_result.to_hdf('L1_integration_summary.hdf', key='data', format='table')
